moved utils.py files into config
