+++ apps/company/views.py
@@ -2,8 +2,8 @@ from django.shortcuts import render
 from rest_framework import viewsets
 from .models import Companies, Branches, BranchBankDetails
 from .serializers import CompaniesSerializer, BranchesSerializer, BranchBankDetailsSerializer
-from utils_methods import list_all_objects, create_instance, update_instance
-from utils_variables import *
+from config.utils_methods import list_all_objects, create_instance, update_instance
+from config.utils_variables import *
 from django_filters.rest_framework import DjangoFilterBackend
 from .filters import CompaniesFilters, BranchesFilters, BranchBankDetailsFilters
 from rest_framework.filters import OrderingFilter
