+++ apps/customer/views.py
@@ -3,7 +3,7 @@ from rest_framework import viewsets
 from apps.customer.filters import LedgerAccountsFilters, CustomerFilters, CustomerAddressesFilters, CustomerAttachmentsFilters
 from .models import *
 from .serializers import *
-from utils_methods import list_all_objects, create_instance, update_instance
+from config.utils_methods import list_all_objects, create_instance, update_instance
 from django_filters.rest_framework import DjangoFilterBackend
 from rest_framework.filters import OrderingFilter
