+++ apps/products/models.py
@@ -2,8 +2,8 @@ import os,uuid
 from django.db import models
 from django.dispatch import receiver
 from django.db.models.signals import pre_delete
-from utils_methods import *
-from utils_variables import *
+from config.utils_methods import *
+from config.utils_variables import *
 from apps.masters.models import ProductUniqueQuantityCodes,ProductTypes,UnitOptions,ProductItemType,ProductDrugTypes,ProductBrands
 
 def product_groups_picture(instance, filename):
