+++ apps/purchase/models.py
@@ -1,11 +1,11 @@
 from django.db import models
-from utils_variables import *
+from config.utils_variables import *
 from apps.masters.models import PurchaseTypes,State,ProductBrands, GstTypes, OrderStatuses
 from apps.customer.models import LedgerAccounts,CustomerCategories
 from apps.vendor.models import Vendor,VendorAgent,VendorAddress,VendorPaymentTerms
 from apps.products.models import Products
 import uuid
-from utils_methods import OrderNumberMixin
+from config.utils_methods import OrderNumberMixin
 
 # Create your models here.
 class PurchaseOrders(OrderNumberMixin):
