+++ config/email_credentials.py
