+++ requirements.txt
@@ -1,41 +1,37 @@
-asgiref==3.7.2
-certifi==2023.11.17
+asgiref==3.8.1
+certifi==2024.6.2
 cffi==1.16.0
+chardet==5.2.0
 charset-normalizer==3.3.2
-colorama==0.4.6
-cryptography==41.0.5
-Django==4.2.7
-django-cors-headers==4.3.0
-django-dotenv==1.4.2
-django-phonenumber-field==7.2.0
-djangorestframework==3.14.0
-djangorestframework-simplejwt==5.3.0
+cryptography==42.0.8
+defusedxml==0.8.0rc2
+Django==5.0.3
+django-cors-headers==4.3.1
+django-templated-mail==1.1.1
+djangorestframework==3.15.1
+djangorestframework-simplejwt==5.3.1
+djoser==2.2.2
 drf-yasg==1.21.7
-idna==3.6
-inflection==0.3.1
-iniconfig==2.0.0
-lazy-object-proxy==1.9.0
-mysqlclient==2.2.0
-packaging==23.2
-phonenumbers==8.13.25
-phonenumberslite==8.13.25
-pluggy==1.3.0
-pycparser==2.21
+idna==3.7
+inflection==0.5.1
+install==1.3.5
+oauthlib==3.2.2
+packaging==24.0
+pillow==10.3.0
+pycparser==2.22
 PyJWT==2.8.0
-PyMySQL==1.1.0
-pytest==7.4.3
-pytest-assert-utils_methods==0.3.1
-pytest-common-subject==1.0.6
-pytest-django==4.7.0
-pytest-drf==1.1.3
-pytest-fixture-order==0.1.4
-pytest-lambda==1.3.0
-python-dotenv==1.0.0
-pytz==2023.3.post1
+PyMySQL==1.1.1
+python-dotenv==1.0.1
+python3-openid==3.2.0
+pytz==2024.1
 PyYAML==6.0.1
-requests==2.31.0
+reportlab==4.2.0
+requests==2.32.3
+requests-oauthlib==2.0.0
+setuptools==70.0.0
+social-auth-app-django==5.4.1
+social-auth-core==4.5.4
 sqlparse==0.4.4
-tzdata==2023.3
+tzdata==2024.1
 uritemplate==4.1.1
-urllib3==2.1.0
-wrapt==1.16.0
+urllib3==2.2.1
